+++ config/index.js
@@ -5,22 +5,25 @@ var config = {
     lng: 'zh_CN',
     avatarDir: process.cwd() + '/../uploads',
     avatarPrefix: '/avatar',
+    mqtt: 'mqtt://root:123@api.h5.jamma.cn',
     modules: {
-      'users': {
+      'user': {
         module: process.cwd() + '/lib'
-      }
+      },
+      'jm-user-mqtt': {}
     }
   },
   production: {
-    port: 3000,
+    port: 80,
     db: 'mongodb://mongo.db/user',
     tokenExpire: 3600,
     disableVerifyCode: false,
     disableAutoUid: false,
     modules: {
-      'users': {
+      'user': {
         module: process.cwd() + '/lib'
-      }
+      },
+      'jm-user-mqtt': {}
     }
   }
 }
@@ -28,5 +31,7 @@ var config = {
 var env = process.env.NODE_ENV || 'development'
 config = config[env] || config['development']
 config.env = env
+
+if (process.env['disableMQTT']) delete config.modules['jm-user-mqtt']
 
 module.exports = config